Mvp-translation/src/translation/requirements.txt
flask
boto3
openai-whisper
torch
ffmpeg-python